% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_add_running_mean.R
\name{add_running_mean}
\alias{add_running_mean}
\title{Running mean for a Hyfe metric}
\usage{
add_running_mean(x, y, window)
}
\arguments{
\item{x}{desc}

\item{y}{desc}

\item{window}{Window size of running mean, in number of values}
}
\value{

}
\description{
Running mean for a Hyfe metric
}
