% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_screening.R
\name{app_screening}
\alias{app_screening}
\title{Shiny app for launching the Hyfe web app based on a participant list.}
\usage{
app_screening(demo = FALSE)
}
\arguments{
\item{demo}{If TRUE, the function will create \code{key.csv} with dummy data
(only if a file named \code{key.csv} does not exist in the working directory)
in order to explore the app's functionality.}
}
\value{
A shiny app.
}
\description{
This app was designed to maximize efficiency in high-throughput screening scenarios.
To work, before this function is called the working directory must have a file named
\code{key.csv}, with columns \code{last} (last name), \code{first} (first name), \code{study.id}
(for the internal ID used to track participants within your study),
\code{consent} (leave blank when file is built --
the app will use this field to track which participants have registered consent),
\code{hyfe.id} (the Hyfe uid), \code{password} (the Hyfe-provided passworded that participants
may use to log into the research app, if the study also involves remote phone monitoring),
and \code{url}, which is the url link to the web app unique for every \code{hyfe.id}.
}
\examples{
app_screening(demo=TRUE)
}
