% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cough_bouts.R
\name{cough_bouts}
\alias{cough_bouts}
\title{Cough bouts}
\usage{
cough_bouts(coughs, bout_window = 2, bout_limit = Inf, verbose = TRUE)
}
\arguments{
\item{coughs}{desc}

\item{bout_window}{desc}

\item{verbose}{desc}
}
\value{

}
\description{
Cough bouts
}
