% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_sessions.R
\name{expand_sessions}
\alias{expand_sessions}
\title{Expand sessions}
\usage{
expand_sessions(
  hyfe_data,
  unit = "hour",
  timestamp_start = NULL,
  timestamp_stop = NULL,
  create_table = TRUE,
  create_series = FALSE,
  inactive_value = 0,
  verbose = FALSE
)
}
\arguments{
\item{hyfe_data}{A hyfe_data object}

\item{unit}{Day, hour (not yet implemented)}

\item{verbose}{Print status updates?}
}
\value{
A list with \code{timetable} (a dataframe with one row per person-date or person-hour)
and \code{series} (a dataframe in which each row is a timestamp, and each column is a UID)
}
\description{
Expand a sessions dataframe (one row per session) into a dataframe with one row per person-date,
with time recorded on that date. This function handles the messiness of
getting the amount of time recorded for a specific person on a specific day
(since a session start on day 1 and ending on day 3 means that all of day 2 was recorded)
}
