% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_find_time_offset.R
\name{find_time_offset}
\alias{find_time_offset}
\title{Background function: find the time offset between Hyfe and a reference}
\usage{
find_time_offset(
  reference_times,
  hyfe_times,
  offset_step,
  offset_range,
  plot_title = NULL,
  toplot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{reference_times}{A vector of numeric timestamps (seconds since 00:00:00 UTC on January 1, 1970).
These are the reference times to which the Hyfe detection times will be compared.
For example, the timestamps of a labelled sound can serve as a reference time.}

\item{hyfe_times}{A vector of numeric timestamps of Hyfe detections.}

\item{offset_step}{The resolution of the offset estimation process; for example, if set to 60,
the function will test various offsets at intervals of 60 seconds.}

\item{offset_range}{The range of potential offsets to evaluate.}

\item{toplot}{If \code{TRUE}, a diagnostic plot will be shown.}

\item{verbose}{Print status updates to the Console?}
}
\value{
A best estimate of Hyfe's time offset, in seconds, from the reference times. The resolution/accuracy of this estimate will depend
on the values chosen for \code{offset_step}.
}
\description{
This function is never called directly by an analyst. It is used in \code{synchronize()}
to determine offset corrections.
}
