% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model_to_cough.R
\name{fit_cough_model}
\alias{fit_cough_model}
\title{Fit mean-variance regression model to cough}
\usage{
fit_cough_model(ho, cutoff_hours = 50, cutoff_hourly = 30, toplot = TRUE)
}
\arguments{
\item{ho}{A \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}

\item{cutoff_hours}{Minimum number of qualifying hours a user must have to be included in the model.}

\item{cutoff_hourly}{Minimum monitoring time within an hour for that hour to be included in the model.}

\item{toplot}{If \code{TRUE}, a regression scatterplot plot will be displayed.}
}
\value{
A list with several named elements: \code{p} is the regression F statistic;
\code{r2} is the adjusted R-squared value for the regression; \code{n} is the number of users included in the
analysis (i.e., the sample size of the model); \code{coefficients} provide the coefficient estimates from
the model (b1 and intercept); \code{model} provids the model object; \code{data} provide the data used to fit the model.
}
\description{
Fit a regression model of the relationship between mean cough rate and its variance,
in which each data point is the mean-variance from a user with sufficient monitoring time.
}
