% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_hyfe_time.R
\name{format_hyfe_time}
\alias{format_hyfe_time}
\title{Format and create date/time variables}
\usage{
format_hyfe_time(timestamps, timezone = NULL)
}
\arguments{
\item{timestamps}{A numeric vector of integer timestamps.
It is assumed that these timestamps represent seconds since 00:00:00 UTC on January 1, 1970,which is the standard format of hyfe_data objects.}

\item{timezone}{A character vector for the timezone, which must match one of the timezones listed in \code{OlsonNames()}.
Most users will retrieve this timezone from \code{hyfe_data$cohort_settings$timezone}.}
}
\value{
A dataframe with various date/time variables. See details \href{https://hyfe-ai.github.io/hyfer/#hyfetime}{here}.
}
\description{
Create a table of standard date/time variables for downstream Hyfe functions.
This function is not usually called by an \code{R} users; it's a background function
called by other \code{hyfer} utilities.
But it can be handy if you have some timestamps that you want to format quickly.
}
\examples{
format_hyfe_time(1626851363)
format_hyfe_time(1626851363, 'UTC')
format_hyfe_time(1626851363, 'Africa/Kampala')
format_hyfe_time(1626851363, 'America/Chicago')

}
