% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ho}
\alias{ho}
\title{Example of a processed \code{hyfe} object (Hyfe test phones) with all users aggregated.
See full details in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}
\format{
An object of class \code{list} of length 10.
}
\usage{
ho
}
\description{
Example of a processed \code{hyfe} object (Hyfe test phones) with all users aggregated.
See full details in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.
}
\keyword{datasets}
