% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ho_by_user}
\alias{ho_by_user}
\title{Example of a processed \code{hyfe} object (Hyfe test phones) with all users processed separately.
See full details in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}
\format{
An object of class \code{list} of length 8.
}
\usage{
ho_by_user
}
\description{
Example of a processed \code{hyfe} object (Hyfe test phones) with all users processed separately.
See full details in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.
}
\keyword{datasets}
