% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hyfe_data}
\alias{hyfe_data}
\title{Example of \code{hyfe_data} object}
\format{
An object of class \code{list} of length 6.
}
\usage{
hyfe_data
}
\description{
Example of \code{hyfe_data} object
}
\keyword{datasets}
