% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyfe_performance.R
\name{hyfe_performance}
\alias{hyfe_performance}
\title{Evaluate Hyfe's performance based on a set of labelled sounds.}
\usage{
hyfe_performance(
  detections,
  labels,
  offsets,
  offset_cutoff = 3,
  prediction_threshold = 0.7,
  n3_cutoff = 10,
  quality_cutoff = 0.95,
  remove_tests = NULL,
  toplot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{detections}{A \code{dataframe} of Hyfe detections. This dataframe requires these columns:
\code{alias} (research/device ID); \code{timestamp} (with a numeric timestamp); \code{n_peaks} (the number of peaks occurring at
the time represented by the timestamp; this can just be 1 if you don't know); and \code{prediction_score} (0-1).}

\item{labels}{A \code{dataframe} with analyst labels. Required columns: \code{test} (numeric ID for each test labeled);
\code{timestamp} (numeric timestamp for the cough-second); \code{label} (best label for this cough-second, 1 - 3, using the Hyfe 4-tier system:
1 = a disputable cough, possibly fake; 2 = definitley a cough but barely audible; 3 = a clear authentic cough);
\code{n1} (if this cough second was labeled category 1 by any analysts, this value is 1; if not it is 0);
same for \code{n2} (category 2) and \code{n3} (category 3);
\code{n_labels} provides the number of labels that occurred within this cough-second;
\code{n_analysts} provides the number of analysts who have reviewed this test;
\code{valid} indicates whether this cough-second should be included in the analysis (\code{TRUE} or \code{FALSE}).}

\item{offsets}{A \code{dataframe} of time offsets, in seconds, for the Hyfe detections in each test represented in the \code{labels} dataset.
Each row is a test. Required fields are: \code{alias} (research/device ID); \code{test} (numeric ID for the test); \code{offset} (seconds
by which to adjust Hyfe timestamp; if postivie, Hyfe timestamps will increase; if negative, they will decrease).}

\item{offset_cutoff}{A number indicating the allowable difference, in seconds, between a Hyfe detection and a label timestamp.}

\item{prediction_threshold}{Cough prediction score threshold.}

\item{n3_cutoff}{Minimum number of category 3 coughs that must be labeled for a test in order for that test to be included in the calculation of performance metrics.}

\item{quality_cutoff}{Minimum quality of a test in order for that test to be included in the estimation of performance metrics. Quality is defined as the fraction of cough-seconds that are category 3.}

\item{remove_tests}{Numeric vector providing the IDs for any tests you want to remove from the performance analysis.}

\item{toplot}{Plot diagnostic plots?}

\item{verbose}{Plot status updates?}
}
\value{
A list with various summaries of the results.
}
\description{
Evaluate Hyfe's performance based on a set of labelled sounds.
}
