% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyfe_summarize.R
\name{hyfe_summarize}
\alias{hyfe_summarize}
\title{Summarize Hyfe data}
\usage{
hyfe_summarize(ho, cutoff_hourly = 30, cutoff_daily = 4)
}
\arguments{
\item{ho}{A \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}

\item{cutoff_hourly}{This function uses sample size cutoffs to ensure that rates are
not swung to extremes due to insufficient monitoring. For example,
an hour of day with 1 cough detection but only 1 minute of monitoring would
produce an hourly cough rate estimate of 60 coughs per hour. Those scenarios should be avoided.
The first cutoff used is \code{cutoff_hourly}: by default, at least 30 minutes of monitoring must occur within a hour-long window of the day
in order for that hour to contribute to the estimation of the hourly cough rate.}

\item{cutoff_daily}{The second cutoff used; the default is that at least 4 hours of monitoring must occur within a day in order
for that day to count toward the daily cough rate.}
}
\value{
A list with two slots: \code{overall}, which summarizes the entire dataset by pooling users together,
and \code{users}, which summarizes each user individually.
The \code{users} slot is \code{NULL} when the \code{hyfe} object provided was processed with \code{by_user = FALSE}.
That \code{users} table, if available, is used to build the \code{overall} slot in which the mean rates (i.e., \code{hourly_rate} and \code{daily_rate}) are the average of each user’s mean rates,
and – importantly – the variability metrics (\code{hourly_var}, \code{hourly_sd}, \code{daily_var}, \code{daily_sd})
now pertain to the variability \emph{among users}.
}
\description{
This function produces summary tables for your \code{hyfe} object.
}
\details{
If your \code{hyfe} object was processed by aggregating all users together
(i.e., your call \code{process_hyfe_data()} included the argument \code{by_user = FALSE},
which is the function's default), the cough rates reported should be treated with caution:
these rates are going to be biased by users with
(1) a lot of monitoring time and (2) a lot of coughs.
To summarize Hyfe data from multiple users in a way that is truly balanced,
in which each user is weighted equally, you should use a \code{hyfe} object processed \verb{with by_user = TRUE}.

Note that the cumulative counts are unaffected by these cutoffs, only the rates.
}
