% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_offset_explorer.R
\name{offset_explorer}
\alias{offset_explorer}
\title{Function for manually finding offsets between a set of labels and Hyfe detections}
\usage{
offset_explorer(alias_id, labels, sounds, prediction_threshold = 0.7)
}
\arguments{
\item{alias_id}{The research alias you want to inspect.}

\item{labels}{A \code{dataframe} with analyst labels. Required columns: \code{test} (numeric ID for each test labeled);
\code{timestamp} (numeric timestamp for the cough-second); and \code{label} (best label for this cough-second, 1 - 3, using the Hyfe 4-tier system:
1 = a disputable cough, possibly fake; 2 = definitley a cough but barely audible; 3 = a clear authentic cough).}

\item{sounds}{The \code{sounds} \code{dataframe} from a \code{hyfe} object.}

\item{prediction_threshold}{Cough prediction score threshold to use for plotting coughs v other noises. Will not change the data.}
}
\value{
Nothing. When this function is run, a Shiny app launches that allows the user to identify the offset. You would
take note of that offset then use it to build up a \code{dataframe} of offsets to pass to \code{hyfe_performance()}.
}
\description{
Function for manually finding offsets between a set of labels and Hyfe detections
}
