% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circadian.R
\name{plot_circadian}
\alias{plot_circadian}
\title{Plot Hyfe circadian patterns}
\usage{
plot_circadian(
  ho,
  type = c("coughs", "sessions", "sounds", "rate"),
  date_min = NULL,
  date_max = NULL,
  by_user = FALSE,
  print_plot = TRUE,
  return_plot = FALSE,
  return_data = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{ho}{desc}

\item{type}{desc}

\item{date_min}{desc}

\item{date_max}{desc}

\item{by_user}{desc}

\item{print_plot}{desc}

\item{return_plot}{desc}

\item{return_data}{desc}

\item{verbose}{desc}

\item{time_unit}{desc}
}
\value{

}
\description{
Plot Hyfe circadian patterns
}
