% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cohort_diagnostic.R
\name{plot_cohort_diagnostic}
\alias{plot_cohort_diagnostic}
\title{Plot diagnostic plot for a cohort}
\usage{
plot_cohort_diagnostic(ho)
}
\arguments{
\item{ho}{desc by user only}
}
\value{

}
\description{
Plot diagnostic plot for a cohort
}
