% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cough_rate_error.R
\name{plot_cough_rate_error}
\alias{plot_cough_rate_error}
\title{Plot cough rate error as a function of monitoring time.}
\usage{
plot_cough_rate_error(hours, cutoff = 0.5, min_error = 0.1, toplot = TRUE)
}
\arguments{
\item{hours}{A dataframe of hourly cough counts. Minimum requirements:
A column \code{coughs} with coughs counted in each hour,
and a column \code{rec} with monitoring time during that hour (0 - 1).
For example, you can use \code{ho$hours} (from a \code{hyfe} object) or the output of \code{simulate_cough_rate()}.}

\item{cutoff}{The amount of monitoring within an hour
in order for that hour to be included in cough rate calculations.
Default is 0.5 (half hour of monitoring during the hour).}

\item{min_error}{Minimum acceptable error. Default is 10\% (0.1).}

\item{toplot}{If \code{TRUE} (the default), a plot will dispaly the result.}
}
\value{
The \code{hours} dataframe with additional columns:
\code{estimate}, the cough rate estimate up to that hour of monitoring;
\code{error}, the error in that estimate based upon the overall cough rate;
\code{prop_error}, the proportional error.
}
\description{
How much monitoring is enough to estimate the overall cough rate accurately?
This is a difficult question to answer, but this plot can help approximate it.
It shows the cough rate estimate based on 1 hour of monitoring, 2 hours, 3 hours, ...,
up until the entire cough record. The plot also displays the 'overall' cough rate, using the entire
monitoring record, and horizontal lines indicating a range of acceptable error around the 'true' rate.
}
