% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_total.R
\name{plot_total}
\alias{plot_total}
\title{Plot cumulative totals over time in your Hyfe dataset}
\usage{
plot_total(
  ho,
  type = c("sessions", "sounds", "coughs"),
  unit = c("days", "hours", "weeks"),
  date_min = NULL,
  date_max = NULL,
  by_user = FALSE,
  print_plot = TRUE,
  return_plot = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{ho}{A \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}

\item{type}{The variable to plot.}

\item{date_min}{Optionally filter the data to a minimum date.
Provide as a character vector of length one, with the format \code{"YYYY-MM-DD HH:MM:SS"}.}

\item{date_max}{Optionally filter to a date maximum. Same format as \code{date_min} above.}

\item{by_user}{If \code{FALSE} (the default), a single line will be plotted that pools all users together.
If \code{TRUE}, a line will be plotted for each user separately.
Note that this is only possible if the call to \code{process_hyfe_data()}
that created the \code{hyfe} object used the argument \code{by_user=TRUE}.}

\item{print_plot}{If \code{TRUE} (the default), the plot will be printed for you.}

\item{return_plot}{If \code{TRUE} (\emph{not} the default), the \code{ggplot} plot object will be returned.
This can be useful if you want to modify/add to the plot (e.g., change axis titles, add a plot title, etc.).}

\item{return_data}{If \code{TRUE} (\emph{not} the default), a simple dataframe will be returned
that provides you with the exact values used to produce the plot.}

\item{time_unit}{The time unit by which to plot it.}
}
\description{
Plot cumulative totals over time in your Hyfe dataset
}
