% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot Hyfe trajectory}
\usage{
plot_trajectory(
  ho,
  type = c("coughs", "sounds", "sessions", "rate"),
  unit = c("days", "hours", "weeks"),
  pool_users = FALSE,
  day_max = NULL,
  print_plot = TRUE,
  return_plot = FALSE,
  return_data = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{ho}{A \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
This function only accepts \code{hyfe} objects that have been processed with \code{ by_user = TRUE}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}

\item{type}{The variable to plot.}

\item{pool_users}{If \code{TRUE}, all user data will be pooled together into a single cumulative line.}

\item{day_max}{Option to control the extent of the X axis. In a trajectory plot, the minimum of the X axis will always be zero.}

\item{print_plot}{If \code{TRUE} (the default), the plot will be printed for you.}

\item{return_plot}{If \code{TRUE} (\emph{not} the default), the \code{ggplot} plot object will be returned.
This can be useful if you want to modify/add to the plot (e.g., change axis titles, add a plot title, etc.).}

\item{return_data}{If \code{TRUE} (\emph{not} the default), a simple dataframe will be returned
that provides you with the exact values used to produce the plot.}

\item{verbose}{Print status updates?}

\item{time_unit}{The time unit by which to plot it.}
}
\value{

}
\description{
In this plot, all users are plotted with the start of their monitoring period beginning at the plot's origin.
This type of plot could be useful if you want to examine patterns across users, such as retention in using the app
or the evoution of cough during a COVID-19 diagnosis.
}
