% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_user_diagnostic.R
\name{plot_user_diagnostic}
\alias{plot_user_diagnostic}
\title{Plot diagnostic plot for a single user}
\usage{
plot_user_diagnostic(ho)
}
\arguments{
\item{ho}{desc}
}
\value{

}
\description{
Plot diagnostic plot for a single user
}
