% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hyfe_data.R
\name{process_hyfe_data}
\alias{process_hyfe_data}
\title{Process a \code{hyfe_data} object for plotting/analysis}
\usage{
process_hyfe_data(hyfe_data, by_user = FALSE, verbose = TRUE)
}
\arguments{
\item{hyfe_data}{A standard \code{hyfe_data} object downloaded
from the Research Dashboard (for external partners) or from \code{hyferdrive} (internal analysts).
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfedata}{package vignette}.}

\item{by_user}{Should data be processed in bulk, or for each user separately?
Processing in bulk (\code{by_user = FALSE}, the default) can be quicker and makes summary plots easier
(e.g., grand total monitoring hours across the cohort). Processing for each user separately (\code{by_user = TRUE})
can take more time at this stage, but opens up more possibilities for plotting and is necessary for unbiased metrics
in which each user is given equal weight (e.g., mean cough rate across the cohort).}

\item{verbose}{Print status updates?}
}
\value{
A \code{hyfe} object, which is a list with several named slots.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.
}
\description{
This function converts a raw \code{hyfe_data} object into a polished \code{hyfe} object.
This function is usually the first step after downloading data.
}
