% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cough_bouts.R
\name{cough_bouts}
\alias{cough_bouts}
\title{Cough bouts}
\usage{
cough_bouts(coughs, bout_window = 2, bout_limit = Inf, verbose = TRUE)
}
\arguments{
\item{coughs}{The \code{coughs} slot from a \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}

\item{bout_window}{The time resolution of a bout. For example, \code{bout_window = 2}
will pool coughs that occur within two seconds of each other into a single cough bout.}

\item{verbose}{Print status updates?}
}
\value{

}
\description{
Cough bouts
}
