% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cough_rate_distribution.R
\name{cough_rate_distribution}
\alias{cough_rate_distribution}
\title{Cough rate distribution}
\usage{
cough_rate_distribution(ho, min_session = 0.5)
}
\arguments{
\item{ho}{desc}

\item{min_session}{desc}
}
\value{

}
\description{
Cough rate distribution
}
