% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cough_rate_distribution.R
\name{cough_rate_distribution}
\alias{cough_rate_distribution}
\title{Cough rate distribution}
\usage{
cough_rate_distribution(ho, min_session = 0.5)
}
\arguments{
\item{ho}{A \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}

\item{min_session}{allows you to define the minimum amount of monitoring required
during a single hour in order for that hour to be included in the cough rate estimation.
For example, sometimes an hour of day contains only a few minutes of monitoring for a user;
that makes for a pretty poor estimate of that hour’s cough rate.
The default \code{min_session} is \code{0.5} hours, or 30 minutes of monitoring within an hour.}
}
\value{
A list with named slots: \code{overall} contains a dataframe with the mean and SD of hourly cough rate for the entire dataset.
These metrics are based on the mean/variance for each individual user, i.e.,
\code{mean_of_mean} is the average of mean cough rates across users.
When using a \code{hyfe} object prepared with \code{by_user=TRUE},
this means that each user is weighted equally in the summary statistics.
When using a \code{hyfe} object in which all user data are aggregated together,
users will be weighted according to their session time;
The slot \verb{$users} contains a dataframe with the mean and SD cough rate for each user.
The slot \verb{$rates} returns a numeric vector of hourly cough rates that satisfy the minimum monitoring threshold:
}
\description{
Summarize cough rates and retrieve hourly cough counts for use in histogram-production and cough distribution modeling.
}
\details{
This function can take both aggregated data (ho) and user-separated (ho_by_user),
but it does best with the latter. It returns metrics about hourly cough rates
based on an hour-by-hour analysis. Similar to the inputs in hyfe_summarize(),
the argument \code{min_session}
The slot \verb{$details} returns a dataframe with all details you might need to analyze these rates (essentially the hours table from a hyfe object).
}
