% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_to_user.R
\name{filter_to_user}
\alias{filter_to_user}
\title{Filter to user}
\usage{
filter_to_user(uid = NULL, hyfe_data)
}
\arguments{
\item{uid}{desc}

\item{hyfe_data}{desc}
}
\value{

}
\description{
Filter to user
}
