% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hyfe_data}
\alias{hyfe_data}
\title{Example of \code{hyfe_data} object
A standard \code{hyfe_data} object downloaded
from the Research Dashboard (for external partners) or from \code{hyferdrive} (internal analysts).
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfedata}{package vignette}.}
\format{
An object of class \code{list} of length 6.
}
\usage{
hyfe_data
}
\description{
Example of \code{hyfe_data} object
A standard \code{hyfe_data} object downloaded
from the Research Dashboard (for external partners) or from \code{hyferdrive} (internal analysts).
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfedata}{package vignette}.
}
\keyword{datasets}
