% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyfe_summarize.R
\name{hyfe_summarize}
\alias{hyfe_summarize}
\title{Summarize Hyfe data}
\usage{
hyfe_summarize(ho, cutoff_hourly = 30, cutoff_daily = 4)
}
\arguments{
\item{ho}{A \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}

\item{cutoff_hourly}{desc}

\item{cutoff_daily}{desc}
}
\value{

}
\description{
Summarize Hyfe data
}
