% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyfe_timetables.R
\name{hyfe_timetables}
\alias{hyfe_timetables}
\title{Produce summary tables of Hyfe data (hourly, daily, weekly)}
\usage{
hyfe_timetables(
  hyfe_data,
  tz = NULL,
  timestamp_start = NULL,
  timestamp_stop = NULL,
  verbose = TRUE
)
}
\arguments{
\item{hyfe_data}{A standard \code{hyfe_data} object downloaded
from the Research Dashboard (for external partners) or from \code{hyferdrive} (internal analysts).
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfedata}{package vignette}.}

\item{tz}{desc}

\item{timestamp_start}{If you want to specify the beginning of the time table, enter timestamp here.}

\item{timestamp_stop}{Specify end of timetable, if you want.}

\item{verbose}{Print status updates?}
}
\value{

}
\description{
Produce summary tables of Hyfe data (hourly, daily, weekly)
}
