% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cohort_diagnostic.R
\name{plot_cohort_diagnostic}
\alias{plot_cohort_diagnostic}
\title{Quick-and-ditry diagnostic plot for a cohort}
\usage{
plot_cohort_diagnostic(ho)
}
\arguments{
\item{ho}{A \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
This function only accepts \code{hyfe} objects that have been processed with \code{by_user = TRUE}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}
}
\value{
A plot with each user's data on a single vertical row.
Grey dots represent monitoring activity, red dots indicate cough detections.
}
\description{
This can be useful in reviewing a cohort dataset or troubleshooting technical issues.
}
