% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot Hyfe timeseries}
\usage{
plot_timeseries(
  ho,
  type = c("coughs", "sounds", "sessions"),
  unit = c("days", "hours", "weeks"),
  date_min = NULL,
  date_max = NULL,
  by_user = FALSE,
  running_mean = NULL,
  print_plot = TRUE,
  return_plot = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{ho}{A \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}

\item{type}{desc}

\item{date_min}{desc}

\item{date_max}{desc}

\item{by_user}{desc}

\item{running_mean}{desc}

\item{print_plot}{desc}

\item{return_plot}{desc}

\item{return_data}{desc}

\item{time_unit}{desc}
}
\value{

}
\description{
Plot Hyfe timeseries
}
