% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot a timeseries of Hyfe counts.}
\usage{
plot_timeseries(
  ho,
  type = c("coughs", "sounds", "sessions"),
  unit = c("days", "hours", "weeks"),
  date_min = NULL,
  date_max = NULL,
  by_user = FALSE,
  running_mean = NULL,
  print_plot = TRUE,
  return_plot = FALSE,
  return_data = FALSE
)
}
\arguments{
\item{ho}{A \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}

\item{type}{The variable to plot.}

\item{date_min}{Optionally filter the data to a minimum date.
Provide as a character vector of length one, with the format \code{"YYYY-MM-DD HH:MM:SS"}.}

\item{date_max}{Optionally filter to a date maximum. Same format as \code{date_min} above.}

\item{by_user}{If \code{FALSE} (the default), a single line will be plotted that pools all users together.
If \code{TRUE}, a line will be plotted for each user separately.
Note that this is only possible if the call to \code{process_hyfe_data()}
that created the \code{hyfe} object used the argument \code{by_user=TRUE}.}

\item{running_mean}{If you wish to add a running mean, provide a binning window here in the units chosen.
This will only work if you are not plotting \code{by_yser}.  If \code{NULL}, no running mean will be plotted.
For example, when \code{time_unit = hours} and \code{running_mean = 24},
the running mean will average the y variable for a 24-hour period, stepping forward one hour at a time.}

\item{print_plot}{If \code{TRUE} (the default), the plot will be printed for you.}

\item{return_plot}{If \code{TRUE} (\emph{not} the default), the \code{ggplot} plot object will be returned.
This can be useful if you want to modify/add to the plot (e.g., change axis titles, add a plot title, etc.).}

\item{return_data}{If \code{TRUE} (\emph{not} the default), a simple dataframe will be returned
that provides you with the exact values used to produce the plot.}

\item{time_unit}{The time unit by which to plot it.}
}
\value{

}
\description{
Plot a timeseries of Hyfe counts.
}
