% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectory.R
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot Hyfe trajectory}
\usage{
plot_trajectory(
  ho,
  type = c("coughs", "sounds", "sessions", "rate"),
  unit = c("days", "hours", "weeks"),
  pool_users = FALSE,
  day_max = NULL,
  print_plot = TRUE,
  return_plot = FALSE,
  return_data = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{ho}{desc - only accepts by user}

\item{type}{desc}

\item{pool_users}{desc}

\item{day_max}{desc}

\item{print_plot}{desc}

\item{return_plot}{desc}

\item{return_data}{desc}

\item{verbose}{desc}

\item{time_unit}{desc}
}
\value{

}
\description{
Plot Hyfe trajectory
}
