% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pool_user_data.R
\name{pool_user_data}
\alias{pool_user_data}
\title{Pool user data}
\usage{
pool_user_data(ho_by_user, group_users = TRUE, verbose = TRUE)
}
\arguments{
\item{ho_by_user}{A \code{hyfe} object, which is generated by \code{process_hyfe_data()}.
This function only accepts \code{hyfe} objects processed with \code{by_user = TRUE}.
See full details and examples in the \href{https://hyfe-ai.github.io/hyfer/#hyfe_object}{package vignette}.}

\item{group_users}{desc}

\item{verbose}{Print status updates?}
}
\value{

}
\description{
Pool user data
}
