% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cough_variance.R
\name{predict_cough_variance}
\alias{predict_cough_variance}
\title{Predict cough variance based on mean-variance regression model}
\usage{
predict_cough_variance(cough_rate, b1, intercept)
}
\arguments{
\item{cough_rate}{The mean cough rate (coughs per hour) for which you want to
predict the variance.}

\item{b1}{The first coefficient of the model.}

\item{intercept}{The second coefficient of the model.}
}
\value{
The predicted variance.
}
\description{
Format of the regression model: (format: variance = exp( b1 *log(mean) +  intercept )).
You can generate these coefficients using the function \code{fit_model_to_cough()}.
}
