% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hyfe_data.R
\name{process_hyfe_data}
\alias{process_hyfe_data}
\title{Process a \code{hyfe_data} object for plotting/analysis}
\usage{
process_hyfe_data(hyfe_data, by_user = FALSE, verbose = TRUE)
}
\arguments{
\item{hyfe_data}{A standard \code{hyfe_data} object.}

\item{by_user}{Should data be processed in bulk, or for each user separately?}

\item{verbose}{Print status updates?}
}
\value{

}
\description{
Process a \code{hyfe_data} object for plotting/analysis
}
